refactor data-analysis related
