+++ libs/common/src/data-analysis/cube-item.interface.ts
