+++ libs/common/src/data-analysis/match-cube-item.interface.ts
