+++ libs/common/src/interfaces/index.ts
@@ -1,3 +1,1 @@
-export * from './cube-item.interface';
-export * from './match-cube-item.interface';
 export * from './player.interface';
