+++ libs/common/src/models/index.ts
@@ -1,7 +1,4 @@
 export * from './common.slack';
 export * from './document';
 export * from './match-result';
-export * from './metrics';
 export * from './player';
-export * from './time-dimension';
-
